auth: reorganize auth service
- reorganized the authentication service from a monolithic
structure to a modern, modular architecture
-  this new organization improves code clarity, simplifies
maintenance, and aligns the service with standard Rust API
project conventions.
