+++ authentication/Cargo.toml
@@ -22,3 +22,4 @@ opentelemetry_sdk = { version = "0.30", features = ["rt-tokio"] }
 opentelemetry-otlp = { version = "0.30", features = ["http-proto", "reqwest-blocking-client"] }
 base64 = "0.22"
 rsa = "0.9"
+thiserror = "1.0"
