+++ authentication/src/handlers/mod.rs
@@ -1,0 +1,4 @@
+pub mod login;
+pub mod register;
+pub mod status;
+pub mod openid;
