+++ authentication/src/state.rs
@@ -1,0 +1,8 @@
+use sqlx::PgPool;
+use crate::config::Config;
+
+#[derive(Clone)]
+pub struct AppState {
+    pub pool: PgPool,
+    pub config: Config,
+}
